refactor: Subtitle to shared
