+++ src/app/components/index.ts
@@ -10,6 +10,5 @@ export { PrincipleCard, type PrincipleCardProps } from './principle-card';
 export { SchoolMenu } from './school-menu';
 export { SectionLabel } from './section-label';
 export { SocialMedia, type SocialMediaProps } from './social-media';
-export { Subtitle } from './subtitle';
 export { TextWithLink } from './text-with-link';
 export { Title, TitleType } from './title';
