+++ src/features/home/contribute/contribute.tsx
@@ -1,8 +1,9 @@
-import { OptionItem, type OptionItemProps, Subtitle, Title } from '@/app/components';
+import { OptionItem, type OptionItemProps, Title } from '@/app/components';
 
 import image from '@/assets/contribute.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { Subtitle } from '@/shared/ui/subtitle';
 
 import './contribute.scss';
