+++ src/features/home/events/events.tsx
@@ -6,12 +6,12 @@ import {
   EventCard,
   EventCardProps,
   SectionLabel,
-  Subtitle,
   Title,
 } from '@/app/components';
 import photo3 from '@/assets/photo-3.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { Subtitle } from '@/shared/ui/subtitle';
 import { getActualDataList } from '@/utils';
 
 import styles from './events.module.scss';
