+++ src/features/home/merch/merch.tsx
@@ -1,4 +1,4 @@
-import { SectionLabel, Subtitle, Title } from '@/app/components';
+import { SectionLabel, Title } from '@/app/components';
 
 import { LINKS } from '@/app/const';
 import image from '@/assets/merch.webp';
@@ -6,6 +6,7 @@ import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { Subtitle } from '@/shared/ui/subtitle';
 
 import './merch.scss';
