+++ src/features/home/numbers/numbers.tsx
@@ -1,7 +1,8 @@
-import { Subtitle, Title } from '@/app/components';
+import { Title } from '@/app/components';
 import image from '@/assets/map.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { Subtitle } from '@/shared/ui/subtitle';
 
 import './numbers.scss';
