+++ src/features/home/support/support.tsx
@@ -1,10 +1,11 @@
-import { Subtitle, Title, TitleType } from '@/app/components';
+import { Title, TitleType } from '@/app/components';
 
 import { LINKS } from '@/app/const';
 import image from '@/assets/support.webp';
 import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
+import { Subtitle } from '@/shared/ui/subtitle';
 
 import './support.scss';
