+++ src/features/school/about/about.tsx
@@ -1,5 +1,6 @@
-import { Subtitle, Title } from '@/app/components';
+import { Title } from '@/app/components';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { Subtitle } from '@/shared/ui/subtitle';
 
 import './about.scss';
