+++ src/shared/ui/subtitle/subtitle.scss
